Create class repository move class entity audit* to repository Entity.Audit
