+++ src/main/java/spring/repository/Audit/AuditClasseRepository.java
@@ -1,0 +1,7 @@
+package spring.repository.Audit;
+
+import org.springframework.data.repository.CrudRepository;
+import spring.entity.Audit.WowAuditClasse;
+
+public interface AuditClasseRepository extends CrudRepository<WowAuditClasse, String> {
+}
