+++ src/main/java/spring/repository/Audit/AuditPersonnageRepository.java
@@ -1,0 +1,7 @@
+package spring.repository.Audit;
+
+import org.springframework.data.repository.CrudRepository;
+import spring.entity.Audit.WowAuditPersonnage;
+
+public interface AuditPersonnageRepository extends CrudRepository<WowAuditPersonnage, String> {
+}
