+++ src/main/java/spring/repository/Audit/AuditProfessionRepository.java
@@ -1,0 +1,7 @@
+package spring.repository.Audit;
+
+import org.springframework.data.repository.CrudRepository;
+import spring.entity.Audit.WowAuditProfession;
+
+public interface AuditProfessionRepository extends CrudRepository<WowAuditProfession, String> {
+}
