+++ src/main/java/spring/repository/Audit/AuditTypePersonnageRepository.java
@@ -1,0 +1,7 @@
+package spring.repository.Audit;
+
+import org.springframework.data.repository.CrudRepository;
+import spring.entity.Audit.WowAuditTypepersonnage;
+
+public interface AuditTypePersonnageRepository extends CrudRepository<WowAuditTypepersonnage, String> {
+}
