+++ src/main/java/spring/repository/Audit/AuditTypeProfessionRepository.java
@@ -1,0 +1,7 @@
+package spring.repository.Audit;
+
+import org.springframework.data.repository.CrudRepository;
+import spring.entity.Audit.WowAuditTypeprofession;
+
+public interface AuditTypeProfessionRepository extends CrudRepository<WowAuditTypeprofession, String> {
+}
