+++ src/main/java/spring/repository/Audit/AuditTypeRaceRepository.java
@@ -1,0 +1,7 @@
+package spring.repository.Audit;
+
+import org.springframework.data.repository.CrudRepository;
+import spring.entity.Audit.WowAuditTyperace;
+
+public interface AuditTypeRaceRepository extends CrudRepository<WowAuditTyperace, String> {
+}
