+++ src/main/java/spring/repository/ClasseRepository.java
@@ -1,0 +1,7 @@
+package spring.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import spring.entity.WowClasse;
+
+public interface ClasseRepository extends CrudRepository<WowClasse, String> {
+}
