+++ src/main/java/spring/repository/PersonnageRepository.java
@@ -1,0 +1,7 @@
+package spring.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import spring.entity.WowPersonnage;
+
+public interface PersonnageRepository extends CrudRepository<WowPersonnage, String> {
+}
