+++ src/main/java/spring/repository/ProfessionRepository.java
@@ -1,0 +1,7 @@
+package spring.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import spring.entity.WowProfession;
+
+public interface ProfessionRepository extends CrudRepository<WowProfession, String> {
+}
