+++ src/main/java/spring/repository/Race.java
@@ -1,0 +1,7 @@
+package spring.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import spring.entity.WowRace;
+
+public interface Race extends CrudRepository<WowRace, String> {
+}
