+++ src/main/java/spring/repository/ServerRepository.java
@@ -1,0 +1,7 @@
+package spring.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import spring.entity.WowServer;
+
+public interface ServerRepository extends CrudRepository<WowServer, String> {
+}
