+++ src/main/java/spring/repository/TypePersonnageRepository.java
@@ -1,0 +1,7 @@
+package spring.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import spring.entity.WowTypePersonnage;
+
+public interface TypePersonnageRepository extends CrudRepository<WowTypePersonnage, String> {
+}
