+++ src/main/java/spring/repository/TypeProfessionRepository.java
@@ -1,0 +1,7 @@
+package spring.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import spring.entity.WowTypeProfession;
+
+public interface TypeProfessionRepository extends CrudRepository<WowTypeProfession, String> {
+}
