+++ src/main/java/spring/repository/TypeRaceRepository.java
@@ -1,0 +1,7 @@
+package spring.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import spring.entity.WowTypeRace;
+
+public interface TypeRaceRepository extends CrudRepository<WowTypeRace, String> {
+}
